requirements.txt
numpy==1.20.1
Pillow==9.1.1
streamlit==1.7.0
